nightshift-53: A2 review-feedback cleanups
Address all four minor suggestions from @claude review on f7c46ac4 +
bbd8a242:

1. Use RecordMetaDataFromProto in the first A2 test (instead of
   rebuilding the schema via the builder) so the records-save path
   uses the EXACT proto bytes that crossed the wire to Java —
   removes any chance of the saved metadata and the record-store
   metadata diverging.

2. Type-assertion failures in the Go index-scan test now return a
   proper error instead of silently producing 0. Surfaces the root
   cause (e.g. tuple-element type mismatch) immediately rather than
   producing misleading "expected 50, got 0" diffs.

3. Java-side sort comment in loadMetaDataAndScanOrdersJava: scoped to
   "future tests with negative ids" — the existing positive-id tests
   don't exercise the negative-tuple-component-ordering case.

4. Java-side prices narrowing: Math.toIntExact(prices.get(i))
   throws ArithmeticException on int32 overflow instead of silently
   wrapping. Constraint is now explicit; current values fit so this
   is purely defensive.

All 5 A2 specs still pass.

+++ conformance/metadata_store_conformance.java
@@ -161,7 +161,7 @@ class MetaDataStoreSteps extends ConformanceBase {
             for (int i = 0; i < orderIds.size(); i++) {
                 Order order = Order.newBuilder()
                     .setOrderId(orderIds.get(i))
-                    .setPrice(prices.get(i).intValue())
+                    .setPrice(Math.toIntExact(prices.get(i)))
                     .build();
                 store.saveRecord(order);
                 saved++;
@@ -234,10 +234,12 @@ class MetaDataStoreSteps extends ConformanceBase {
                     throw new RuntimeException("Failed to parse Order: " + e.getMessage(), e);
                 }
             }).join();
-            // Sort by orderId for deterministic Go-side comparison —
-            // scanRecords is PK-ordered already, but record-layer sorts
-            // tuples by component, not necessarily by long-int value alone
-            // when negatives are involved; explicit sort is harmless.
+            // Defensive sort by orderId for deterministic Go-side
+            // comparison. scanRecords is PK-ordered already and current
+            // tests use positive ids, so this is a no-op today; the
+            // explicit sort is a safety net for future tests that might
+            // use negative ids (where tuple-component ordering would
+            // surface ahead of strict numeric ordering).
             rows.sort((a, b) -> Long.compare((Long) a.get("orderId"), (Long) b.get("orderId")));
             result.put("found", true);
             result.put("metadataVersion", proto.getVersion());
@@ -289,7 +291,7 @@ class MetaDataStoreSteps extends ConformanceBase {
             for (int i = 0; i < orderIds.size(); i++) {
                 Order order = Order.newBuilder()
                     .setOrderId(orderIds.get(i))
-                    .setPrice(prices.get(i).intValue())
+                    .setPrice(Math.toIntExact(prices.get(i)))
                     .build();
                 store.saveRecord(order);
             }
